+++ lagom/markers.py
@@ -1,5 +1,7 @@
 from typing import Any
 
+from lagom.exceptions import InjectableMarkerConsumed
+
 
 class Injectable:
     """
@@ -14,6 +16,9 @@ class Injectable:
         :return:
         """
         return False
+
+    def __getattribute__(self, item):
+        raise InjectableMarkerConsumed()
 
 
 # singleton object used to indicate that an argument should be injected